+++ clang/lib/Parse/ParseDeclCXX.cpp
@@ -2252,9 +2252,6 @@ void Parser::ParseBaseClause(Decl *ClassDecl) {
   while (true) {
     // Parse a base-specifier.
     BaseResult Result = ParseBaseSpecifier(ClassDecl);
-    // Skip any base-specifier we couldn’t actually build into a usable
-    // CXXBaseSpecifier (covers both syntactic invalidity and
-    // other un-usable cases).
     if (!Result.isUsable()) {
       // Skip the rest of this base specifier, up until the comma or
       // opening brace.