+++ clang/test/SemaCXX/invalid-base-inheritance.cpp
@@ -1,6 +1,8 @@
 // Tests that invalid base-specifiers no longer crash the compiler.
 // RUN: %clang_cc1 -std=c++20 -fsyntax-only -verify %s
 
+namespace GH147186 {
+
 class X; // expected-note {{forward declaration of 'X'}} expected-note {{forward declaration of 'X'}}
 
 class A : X { // expected-error {{base class has incomplete type}}
@@ -11,3 +13,4 @@ class Y : int { // expected-error {{expected class name}}
 
 class Z : X*, virtual int { // expected-error {{base class has incomplete type}} expected-error {{expected class name}}
 };
+}